refactor

+++ src/features/posts/api/posts.controller.ts
@@ -9,6 +9,7 @@ import {
   Query,
   Put,
   Delete,
+  HttpException,
 } from '@nestjs/common';
 import { PostsService } from '../application/posts.service';
 import { PostsQueryRepo } from '../infrastructure/posts.query-repo';
@@ -17,12 +18,14 @@ import { PostsQueryParamsInputModel } from './models/input/posts-query-params.in
 
 import { UpdatePostInputModel } from './models/input/update-post.input.model';
 import { setPagination } from '../../../utils/set-pagination';
+import { PostsRepo } from '../infrastructure/posts.repo';
 
 @Controller('posts')
 export class PostsController {
   constructor(
     private postsService: PostsService,
-    private postsQueryRepo: PostsQueryRepo
+    private postsQueryRepo: PostsQueryRepo,
+    private postsRepo: PostsRepo
   ) {}
   @Get()
   async getAll(
@@ -34,6 +37,11 @@ export class PostsController {
 
   @Get(':id')
   async getBlogById(@Param('id') id: string) {
+    const existsBlog = await this.postsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Post with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     return await this.postsQueryRepo.findById(id);
   }
 
@@ -47,12 +55,22 @@ export class PostsController {
   @Put(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async updatePost(@Param('id') id: string, @Body() updatePostModel: UpdatePostInputModel) {
+    const existsBlog = await this.postsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Post with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.postsService.updatePost(id, updatePostModel);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async deletePost(@Param('id') id: string) {
+    const existsBlog = await this.postsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Post with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.postsService.deletePost(id);
   }
 }
